🎨 To be reverted see commit msg
The style of the listing of the nested groups on the interface page is suboptimal since it is made for displaying code. This commit contains some css code that would enable better styling IF one would be able to add HTML tag with the class qdmi_interface as a sibling into those HTML elements.