✨ Introducing Duration Unit and Reporting Shuttling Speed (#210)
## Description

This PR adds properties to report the duration unit used by the device,
similarly to the length unit introduced in #198.

On top, it adds the feature to report the shuttling speed requested in
#191.

Fixes #191

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.