📝 Improve Docstrings related to Length Values (#211)
## Description

In #210, the duration unit was added, similar to the length unit that
was already added previously in #198. #210 provided a better formulation
of the corresponding docstring for duration related properties. This PR
mirrors this better formulation and adapts it for all length related
properties.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.