Fix bug in example tool (#218)
## Description

There was a bug in `Tool::compile`. The original code could never
(correctly) replace the qubits (e.g. `q[0]` by `q[<replaced id>]`), due
to two subtle issues

- Clearing the stringstream did not actually empty it (see e.g.
`from.clear()`), hence it accumulated garbage preventing the pattern
matching to ever match the qubits.
- `edge.first` and `edge.second` are just pointer and hence print to but
numbers and not the desired ids of the sites.

This PR fixes those things. There is a test for `Tool::compile` (see
`TEST_P(QDMIImplementationTest, ToolCompile)`) but it could not capture
this bug because the input and expected output are mostly the same which
is replicated by doing (mostly) nothing.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [ ] I have added appropriate tests that cover the new/changed
functionality.
- [ ] I have updated the documentation to reflect these changes.
- [ ] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or
      removals.
- [ ] I have added migration instructions to the upgrade guide (if
needed).
- [ ] The changes follow the project's style guidelines and introduce no
new warnings.
- [ ] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Co-authored-by: Reinhard Stahn <r.stahn@parityqc.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>