👷🏼 Better support for PRs from forks (#219)
## Description

This PR aims to improve the developer experience for PRs from forks,
which would previously fail due to a lack of permissions.
The docs preview is now no longer being deployed on PRs from forks.
Some extra permissions are given to the testing workflow. Although I
have to say that I am not sure what is going on with codecov. The exact
same setup is working nicely over at the MQT.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or
      removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>