✨ Update CMake config and External Dependencies (#250)
## Description

This PR updates the general CMake configuration of the project.
It changes the minimum required CMake version to 3.24, which allows
simplifying some of the FetchContent usage.
Additionally, it updates all external dependencies, including doxygen,
which allows us to fix the long-standing issues observed in #146 and
initially tackled in #148.

Fixes #146 

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or
      removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>