🐛 Fix target name mismatch for `qdmi_project_warnings` between source and installed version of QDMI (#270)
## Description

This fixes an oversight in the QDMI installation instructions that would
lead to a target mismatch between a source version of QDMI and an
installed version.
The source version would refer to `qdmi::project_warnings`, while the
installed version provides `qdmi::qdmi_project_warnings`.
The main reason for that is that the installation instructions prefix
all regular target names with `qdmi::` (which is intended).
The reason this hasn't been caught in the `USE_INSTALLED_QDMI` CI tests
is that the main `CMakeLists.txt` checked whether it could find the
`qdmi::project_warnings` target and define it anew if it could not.
While this works locally in the QDMI repository, this does not work in
an external project that tries to use an installed version of QDMI (that
version won't run through the original `CMakeLists.txt`, but only
through the package config file `qdmi-config.cmake`).

The fix here is to adjust the respective target alias and to change all
occurrences of `qdmi::project_warnings` in the codebase to
`qdmi::qdmi_project_warnings`.

Note that we cannot simply call the target `project_warnings` as this
has a high risk of conflicting with other CMake-based projects such as
the MQT. Thus, the `qdmi` duplication in the prefixed target name is
kind of necessary.

Any consuming project will have to update their link target if they
previously linked against `qdmi::project_warnings`.
I will backport this to a `v1.2.x` branch so that we can release a patch
version.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or
      removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>
(cherry picked from commit 3c8e1b13272fab1c2bb478456fb5d713add45bcb)