🚸 Provide compile definitions for inferring device and QDMI versions automatically (#272)
## Description

This PR adds some compile definitions to the example device as well as
the template that allow to automatically infer the version in the device
implementation from the versions set through CMake.
One less place to update the version of QDMI for us (and everyone else).

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or
      removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>

(cherry picked from commit d11411ea7aa918aa3aaf9f02b689be0e567cbbd2)
Signed-off-by: burgholzer <burgholzer@me.com>