🐛 Ensure devices can be instantiated from an installed version of QDMI (#273)
## Description

This is another follow-up pull request that fixes usage patterns when
relying on an installed version of QDMI.
In particular, it moves the `test_defs.cpp.in` file from the
`test/utils` to the `cmake` directory and installs it as part of the
package.
Additionally, it makes sure that the PrefixHandling functions can be
sensibly used from an installed version of QDMI.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or
      removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>
(cherry picked from commit 6caddb0b90afddfd4cef432c66fa842ba27ea012)